==> rna/R_Code/02_calculate_dcor.ipynb <==
# Setup -------------------------------------------------------------------
# Importiere Pakete
suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(Rfast))
suppressPackageStartupMessages(library(Rcpp))
suppressPackageStartupMessages(library(RcppArmadillo))

# Userinputs --------------------------------------------------------------

filename_rawdata <- "../data/agg_gene_data/agg_gene_data_short_ALL.csv"

# Lese benoetigte Daten ein -----------------------------------------------
geneData <- read.csv(file = filename_rawdata)

# Entferne Spalten mit Zeilenindizes und Sample-ID
geneMetadata <- geneData[,c("id", "DIAGNOSIS")]
geneData[, c("X", "id", "DIAGNOSIS")] <- NULL

head(geneMetadata)

message("Anzahl vorhandener Gene:")
(n_correlation_rows <- ncol(geneData))

# Importiere kompilierten C-Code
sourceCpp("../C_Code/my_dcor.cpp")

# Allokiere Korrelationsmatrix
corr_matrix <- matrix(0,
                      nrow = n_correlation_rows,
                      ncol = n_correlation_rows)

# Wandel Daten vom Dataframe in eine Matrix 
X <- data.matrix(geneData)

corr_matrix_c <- my_dcor(X, logfile = "/home/jovyan/my_log.txt") # Fix path to log-file!

# Sicherung der Korrelationsmatrix ----------------------------------------

# Dateiname im Format "corr_matrix_1000_2021-11-17.csv"
filename_corrmatrix <- paste0("../data/correlation_matrices/corr_matrix_",
                              n_correlation_rows, "_",
                              Sys.Date(), ".csv")
                              
filename_metadata <- paste0("../data/correlation_matrices/corr_matrix_",
                              n_correlation_rows, "_",
                              Sys.Date(), "metadata.csv")

# Sichere berechnete Matrix in .csv-Datei
message("Sichere Korrelationsmatrix unter: ", filename_corrmatrix)
write.table(corr_matrix_c,
            file = filename_corrmatrix,
            row.names = FALSE,
            col.names = FALSE)
            
write.table(geneMetadata,
            file = filename_metadata,
            row.names = FALSE,
            col.names = FALSE)
